Centralize validation utilities
