+++ src/main/java/com/amannmalik/mcp/McpServer.java
@@ -342,13 +342,6 @@ public final class McpServer implements AutoCloseable {
         }
     }
 
-    private JsonRpcError invalidParams(JsonRpcRequest req, String message) {
-        return JsonRpcError.invalidParams(req.id(), message);
-    }
-
-    private JsonRpcError invalidParams(JsonRpcRequest req, IllegalArgumentException e) {
-        return invalidParams(req, e.getMessage());
-    }
 
     private Optional<String> rateLimit(RateLimiter limiter, String key) {
         try {
@@ -377,25 +370,16 @@ public final class McpServer implements AutoCloseable {
 
     private JsonRpcMessage listTools(JsonRpcRequest req) {
         requireServerCapability(ServerCapability.TOOLS);
-        try {
-            ListToolsRequest ltr = ListToolsRequest.CODEC.fromJson(req.params());
-            String cursor = sanitizeCursor(ltr.cursor());
-            Pagination.Page<Tool> page = tools.list(cursor);
-            JsonObject json = ListToolsResult.CODEC.toJson(new ListToolsResult(page.items(), page.nextCursor(), null));
-            return new JsonRpcResponse(req.id(), json);
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e.getMessage());
-        }
+        ListToolsRequest ltr = ListToolsRequest.CODEC.fromJson(req.params());
+        String cursor = sanitizeCursor(ltr.cursor());
+        Pagination.Page<Tool> page = tools.list(cursor);
+        JsonObject json = ListToolsResult.CODEC.toJson(new ListToolsResult(page.items(), page.nextCursor(), null));
+        return new JsonRpcResponse(req.id(), json);
     }
 
     private JsonRpcMessage callTool(JsonRpcRequest req) {
         requireServerCapability(ServerCapability.TOOLS);
-        CallToolRequest callRequest;
-        try {
-            callRequest = CallToolRequest.CODEC.fromJson(req.params());
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e);
-        }
+        CallToolRequest callRequest = CallToolRequest.CODEC.fromJson(req.params());
         Optional<String> limit = rateLimit(toolLimiter, callRequest.name());
         if (limit.isPresent()) {
             return JsonRpcError.of(req.id(), RATE_LIMIT_CODE, limit.get());
@@ -409,72 +393,48 @@ public final class McpServer implements AutoCloseable {
             ToolResult result = tools.call(callRequest.name(), callRequest.arguments());
             return new JsonRpcResponse(req.id(), ToolResult.CODEC.toJson(result));
         } catch (IllegalArgumentException e) {
-            return handleToolCallFailure(req, callRequest, e);
-        }
-    }
-
-    private JsonRpcMessage handleToolCallFailure(JsonRpcRequest req, CallToolRequest callRequest, IllegalArgumentException e) {
-        Optional<Tool> tool = tools.find(callRequest.name());
-        if (tool.isPresent() && lifecycle.negotiatedClientCapabilities().contains(ClientCapability.ELICITATION)) {
-            try {
-                ElicitRequest er = new ElicitRequest(
-                        "Provide arguments for tool '" + tool.get().name() + "'",
-                        tool.get().inputSchema(),
-                        null);
-                ElicitResult res = elicit(er);
-                if (res.action() == ElicitationAction.ACCEPT) {
-                    try {
+            Optional<Tool> tool = tools.find(callRequest.name());
+            if (tool.isPresent() && lifecycle.negotiatedClientCapabilities().contains(ClientCapability.ELICITATION)) {
+                try {
+                    ElicitRequest er = new ElicitRequest(
+                            "Provide arguments for tool '" + tool.get().name() + "'",
+                            tool.get().inputSchema(),
+                            null);
+                    ElicitResult res = elicit(er);
+                    if (res.action() == ElicitationAction.ACCEPT) {
                         ToolResult result = tools.call(callRequest.name(), res.content());
                         return new JsonRpcResponse(req.id(), ToolResult.CODEC.toJson(result));
-                    } catch (IllegalArgumentException ex) {
-                        return invalidParams(req, ex);
                     }
+                    throw new IllegalArgumentException("Tool invocation cancelled");
+                } catch (IOException io) {
+                    return JsonRpcError.of(req.id(), JsonRpcErrorCode.INTERNAL_ERROR, io.getMessage());
                 }
-                return invalidParams(req, "Tool invocation cancelled");
-            } catch (IllegalArgumentException ex) {
-                return invalidParams(req, ex);
-            } catch (Exception ex) {
-                return JsonRpcError.of(req.id(), JsonRpcErrorCode.INTERNAL_ERROR, ex.getMessage());
             }
+            throw e;
         }
-        return invalidParams(req, e);
     }
 
     private JsonRpcMessage listPrompts(JsonRpcRequest req) {
         requireServerCapability(ServerCapability.PROMPTS);
-        try {
-            ListPromptsRequest lpr = ListPromptsRequest.CODEC.fromJson(req.params());
-            String cursor = sanitizeCursor(lpr.cursor());
-            Pagination.Page<Prompt> page = prompts.list(cursor);
-            return new JsonRpcResponse(req.id(), ListPromptsResult.CODEC.toJson(new ListPromptsResult(page.items(), page.nextCursor(), null)));
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e.getMessage());
-        }
+        ListPromptsRequest lpr = ListPromptsRequest.CODEC.fromJson(req.params());
+        String cursor = sanitizeCursor(lpr.cursor());
+        Pagination.Page<Prompt> page = prompts.list(cursor);
+        return new JsonRpcResponse(req.id(), ListPromptsResult.CODEC.toJson(new ListPromptsResult(page.items(), page.nextCursor(), null)));
     }
 
     private JsonRpcMessage getPrompt(JsonRpcRequest req) {
         requireServerCapability(ServerCapability.PROMPTS);
-        try {
-            GetPromptRequest getRequest = GetPromptRequest.CODEC.fromJson(req.params());
-            PromptInstance inst = prompts.get(getRequest.name(), getRequest.arguments());
-            return new JsonRpcResponse(req.id(), PromptInstance.CODEC.toJson(inst));
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e.getMessage());
-        }
+        GetPromptRequest getRequest = GetPromptRequest.CODEC.fromJson(req.params());
+        PromptInstance inst = prompts.get(getRequest.name(), getRequest.arguments());
+        return new JsonRpcResponse(req.id(), PromptInstance.CODEC.toJson(inst));
     }
 
     private JsonRpcMessage setLogLevel(JsonRpcRequest req) {
         requireServerCapability(ServerCapability.LOGGING);
         JsonObject params = req.params();
-        if (params == null) {
-            return invalidParams(req, "Missing params");
-        }
-        try {
-            logLevel = SetLevelRequest.CODEC.fromJson(params).level();
-            return new JsonRpcResponse(req.id(), JsonValue.EMPTY_JSON_OBJECT);
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e.getMessage());
-        }
+        if (params == null) throw new IllegalArgumentException("Missing params");
+        logLevel = SetLevelRequest.CODEC.fromJson(params).level();
+        return new JsonRpcResponse(req.id(), JsonValue.EMPTY_JSON_OBJECT);
     }
 
     private void sendLog(LoggingMessageNotification note) throws IOException {
@@ -495,19 +455,15 @@ public final class McpServer implements AutoCloseable {
         }
         requireServerCapability(ServerCapability.COMPLETIONS);
         JsonObject params = req.params();
-        if (params == null) {
-            return invalidParams(req, "Missing params");
+        if (params == null) throw new IllegalArgumentException("Missing params");
+        CompleteRequest request = CompleteRequest.CODEC.fromJson(params);
+        Optional<String> limit = rateLimit(completionLimiter, request.ref().toString());
+        if (limit.isPresent()) {
+            return JsonRpcError.of(req.id(), RATE_LIMIT_CODE, limit.get());
         }
         try {
-            CompleteRequest request = CompleteRequest.CODEC.fromJson(params);
-            Optional<String> limit = rateLimit(completionLimiter, request.ref().toString());
-            if (limit.isPresent()) {
-                return JsonRpcError.of(req.id(), RATE_LIMIT_CODE, limit.get());
-            }
             CompleteResult result = completions.complete(request);
             return new JsonRpcResponse(req.id(), CompleteResult.CODEC.toJson(result));
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e.getMessage());
         } catch (Exception e) {
             return JsonRpcError.of(req.id(), JsonRpcErrorCode.INTERNAL_ERROR, e.getMessage());
         }
@@ -595,15 +551,11 @@ public final class McpServer implements AutoCloseable {
 
     private JsonRpcMessage handleCreateMessage(JsonRpcRequest req) {
         JsonObject params = req.params();
-        if (params == null) {
-            return invalidParams(req, "Missing params");
-        }
+        if (params == null) throw new IllegalArgumentException("Missing params");
         try {
             CreateMessageRequest cmr = CreateMessageRequest.CODEC.fromJson(params);
             CreateMessageResponse resp = createMessage(cmr);
             return new JsonRpcResponse(req.id(), CreateMessageResponse.CODEC.toJson(resp));
-        } catch (IllegalArgumentException e) {
-            return invalidParams(req, e.getMessage());
         } catch (Exception e) {
             return JsonRpcError.of(req.id(), JsonRpcErrorCode.INTERNAL_ERROR, e.getMessage());
         }
